Migrate from raven to sentry-go package
Adds a sentry subpackage to the project, which interfaces
with Sentry via the sentry-go package instead of using
the legacy go raven client. The old raven subpackage has
been kept in order to more easily facilitate a migration
between the two packages.

In order to migrate use of the legacy raven client,
calls to raven.CaptureErrors and CaptureErrorsAltDsn
should be replaced with the singular sentry.CaptureErrors.
Use of raven.AltDsn and raven.Fingerprint should be
replaced with equivalent use of sentry.AltDsn and
sentry.Fingerprint.

Rewrites the stacktrace module to make use of the
sentry-go package Stacktrace struct. Moves the stacktrace
files used by the raven client to a subdirectory of the
raven folder. In the new sentry subpackage, equivalent
tests exist in the sentry-go package since we are now
piggybacking on their stacktrace building code.

Adds a robust test that verifies the output of the
events sent to Sentry for glog.Error calls, including
differences between passing strings, raw error objects,
or xerror/yerror objects to glog. Also adds a test in
the stacktrace subpackage which verifies paths are
normalized properly.

J=SRE-3390