sentry: remove unique error identifiers to improve event grouping
Changes the Type and Value fields of exceptions sent to Sentry
such that, whenever possible, any unique components of the
error text such as IDs are removed from the Type field.
The Type field is utilized in the Sentry event-merge
algorithm in comparing stacktraces, so if the Type field
of two issues differs the stacktraces are not considered
identical, resulting in a duplicate issue appearing in
Sentry.

The Type field is now split on the ":" character, if it
exists, with the part before the : appearing as the type
and the part after appearing as the Value. If a glog
event is captured directly via glog.Errorf, then the
glog.FormatStringArg is captured so that the raw
format string (containing %s and other formatters)
can be processed and set as the Type, with the Value
field containing the remainder of the arguments.

J=SRE-3390
TEST=auto
  Updated backend_test.go