sentry: apply existing scope
Previously, the glog integration did
not respect existing scopes set by the
sentry-go package within an application.

This is because glog was creating new hubs
and scopes per dsn so it could dynamically
route glog Errors to the proper team so it
wasn't inheriting the default scope automatically.

This change updates the integration to apply the
current scope to the event which will allow
teams to persist relevant data to sentry from
any caller of a glog.Error function.

TEST=auto,manual

Added unit tests. Created a program
which configured sentry locally
with a valid DSN and confirmed the additional
context appeared in sentry's platform.